requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
dotenv==0.9.9
fastapi==0.124.1
idna==3.11
mysql-connector==2.2.9
mysql-connector-python==9.5.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
